Split event definitions by domain
