+++ pkg/events/chat-events.go
@@ -136,234 +136,6 @@ func (e *EventImpl) SetPayload(b []byte) {
 
 var _ Event = &EventImpl{}
 
-type EventPartialCompletionStart struct {
-	EventImpl
-}
-
-func NewStartEvent(metadata EventMetadata) *EventPartialCompletionStart {
-	return &EventPartialCompletionStart{
-		EventImpl: EventImpl{
-			Type_:     EventTypeStart,
-			Metadata_: metadata,
-			payload:   nil,
-		},
-	}
-}
-
-var _ Event = &EventPartialCompletionStart{}
-
-type EventInterrupt struct {
-	EventImpl
-	Text string `json:"text"`
-	// TODO(manuel, 2024-07-04) Add all collected tool calls so far
-}
-
-func NewInterruptEvent(metadata EventMetadata, text string) *EventInterrupt {
-	return &EventInterrupt{
-		EventImpl: EventImpl{
-			Type_:     EventTypeInterrupt,
-			Metadata_: metadata,
-			payload:   nil,
-		},
-		Text: text,
-	}
-}
-
-var _ Event = &EventInterrupt{}
-
-type EventFinal struct {
-	EventImpl
-	Text string `json:"text"`
-	// TODO(manuel, 2024-07-04) Add all collected tool calls so far
-}
-
-func NewFinalEvent(metadata EventMetadata, text string) *EventFinal {
-	return &EventFinal{
-		EventImpl: EventImpl{
-			Type_:     EventTypeFinal,
-			Metadata_: metadata,
-			payload:   nil,
-		},
-		Text: text,
-	}
-}
-
-var _ Event = &EventFinal{}
-
-type EventError struct {
-	EventImpl
-	ErrorString string `json:"error_string"`
-}
-
-func NewErrorEvent(metadata EventMetadata, err error) *EventError {
-	return &EventError{
-		EventImpl: EventImpl{
-			Type_:     EventTypeError,
-			Metadata_: metadata,
-			payload:   nil,
-		},
-		ErrorString: err.Error(),
-	}
-}
-
-var _ Event = &EventError{}
-
-// TODO(manuel, 2024-07-05) This might be possible to delete
-type EventText struct {
-	EventImpl
-	Text string `json:"text"`
-	// TODO(manuel, 2024-06-04) Add ToolCall information here, and potentially multiple responses (see the claude API that allows multiple content blocks)
-	// This is currently stored in the metadata uder the MetadataToolCallsSlug (see chat-with-tools-step.go in openai)
-}
-
-func NewTextEvent(metadata EventMetadata, text string) *EventText {
-	return &EventText{
-		EventImpl: EventImpl{
-			Type_:     EventTypeStart,
-			Metadata_: metadata,
-			payload:   nil,
-		},
-		Text: text,
-	}
-}
-
-var _ Event = &EventText{}
-
-type ToolCall struct {
-	ID    string `json:"id"`
-	Name  string `json:"name"`
-	Input string `json:"input"`
-}
-
-// TODO(manuel, 2024-07-04) Handle multiple tool calls
-type EventToolCall struct {
-	EventImpl
-	ToolCall ToolCall `json:"tool_call"`
-}
-
-func NewToolCallEvent(metadata EventMetadata, toolCall ToolCall) *EventToolCall {
-	return &EventToolCall{
-		EventImpl: EventImpl{
-			Type_:     EventTypeToolCall,
-			Metadata_: metadata,
-			payload:   nil,
-		},
-		ToolCall: toolCall,
-	}
-}
-
-var _ Event = &EventToolCall{}
-
-type ToolResult struct {
-	ID     string `json:"id"`
-	Name   string `json:"name,omitempty"`
-	Result string `json:"result"`
-}
-
-type EventToolResult struct {
-	EventImpl
-	ToolResult ToolResult `json:"tool_result"`
-}
-
-func NewToolResultEvent(metadata EventMetadata, toolResult ToolResult) *EventToolResult {
-	return &EventToolResult{
-		EventImpl: EventImpl{
-			Type_:     EventTypeToolResult,
-			Metadata_: metadata,
-			payload:   nil,
-		},
-		ToolResult: toolResult,
-	}
-}
-
-var _ Event = &EventToolResult{}
-
-// EventToolCallExecute captures the intent to execute a tool locally
-type EventToolCallExecute struct {
-	EventImpl
-	ToolCall ToolCall `json:"tool_call"`
-}
-
-func NewToolCallExecuteEvent(metadata EventMetadata, toolCall ToolCall) *EventToolCallExecute {
-	return &EventToolCallExecute{
-		EventImpl: EventImpl{
-			Type_:     EventTypeToolCallExecute,
-			Metadata_: metadata,
-			payload:   nil,
-		},
-		ToolCall: toolCall,
-	}
-}
-
-var _ Event = &EventToolCallExecute{}
-
-// EventToolCallExecutionResult captures the result of executing a tool locally
-type EventToolCallExecutionResult struct {
-	EventImpl
-	ToolResult ToolResult `json:"tool_result"`
-}
-
-func NewToolCallExecutionResultEvent(metadata EventMetadata, toolResult ToolResult) *EventToolCallExecutionResult {
-	return &EventToolCallExecutionResult{
-		EventImpl: EventImpl{
-			Type_:     EventTypeToolCallExecutionResult,
-			Metadata_: metadata,
-			payload:   nil,
-		},
-		ToolResult: toolResult,
-	}
-}
-
-var _ Event = &EventToolCallExecutionResult{}
-
-// TODO(manuel, 2024-07-03) Then, we can add those to the openai step as well, and to the UI, and then we should have a good way to do auto tool calling
-
-// EventPartialCompletion is the event type for textual partial completion. We don't support partial tool completion.
-type EventPartialCompletion struct {
-	EventImpl
-	Delta string `json:"delta"`
-	// This is the complete completion string so far (when using openai, this is currently also the toolcall json)
-	Completion string `json:"completion"`
-
-	// TODO(manuel, 2024-06-04) This might need partial tool completion if it is of interest,
-	// this is less important than adding tool call information to the result above
-}
-
-func NewPartialCompletionEvent(metadata EventMetadata, delta string, completion string) *EventPartialCompletion {
-	return &EventPartialCompletion{
-		EventImpl: EventImpl{
-			Type_:     EventTypePartialCompletion,
-			Metadata_: metadata,
-			payload:   nil,
-		},
-		Delta:      delta,
-		Completion: completion,
-	}
-}
-
-var _ Event = &EventPartialCompletion{}
-
-// EventThinkingPartial mirrors EventPartialCompletion but is dedicated to reasoning/summary text
-type EventThinkingPartial struct {
-	EventImpl
-	Delta      string `json:"delta"`
-	Completion string `json:"completion"`
-}
-
-func NewThinkingPartialEvent(metadata EventMetadata, delta string, completion string) *EventThinkingPartial {
-	return &EventThinkingPartial{
-		EventImpl:  EventImpl{Type_: EventTypePartialThinking, Metadata_: metadata},
-		Delta:      delta,
-		Completion: completion,
-	}
-}
-
-var _ Event = &EventThinkingPartial{}
-
-// MetadataToolCallsSlug is the slug used to store ToolCall metadata as returned by the openai API
-// TODO(manuel, 2024-07-04) This needs to deleted once we have a good way to do tool calling
-const MetadataToolCallsSlug = "tool-calls"
-
 // EventMetadata contains all the information that is passed along with watermill message,
 // specific to chat steps.
 type EventMetadata struct {
@@ -741,399 +513,4 @@ func (e *EventImpl) ToToolCall() (EventToolCall, bool) {
 		return EventToolCall{}, false
 	}
 	return *ret, true
-}
-
-func (e EventPartialCompletionStart) MarshalZerologObject(ev *zerolog.Event) {
-	e.EventImpl.MarshalZerologObject(ev)
-}
-
-func (e EventInterrupt) MarshalZerologObject(ev *zerolog.Event) {
-	e.EventImpl.MarshalZerologObject(ev)
-	ev.Str("text", e.Text)
-}
-
-func (e EventFinal) MarshalZerologObject(ev *zerolog.Event) {
-	e.EventImpl.MarshalZerologObject(ev)
-	ev.Str("text", e.Text)
-}
-
-func (e EventError) MarshalZerologObject(ev *zerolog.Event) {
-	e.EventImpl.MarshalZerologObject(ev)
-	ev.Str("error", e.ErrorString)
-}
-
-func (e EventText) MarshalZerologObject(ev *zerolog.Event) {
-	e.EventImpl.MarshalZerologObject(ev)
-	ev.Str("text", e.Text)
-}
-
-func (tc ToolCall) MarshalZerologObject(ev *zerolog.Event) {
-	ev.Str("id", tc.ID).Str("name", tc.Name).Str("input", tc.Input)
-}
-
-func (e EventToolCall) MarshalZerologObject(ev *zerolog.Event) {
-	e.EventImpl.MarshalZerologObject(ev)
-	ev.Object("tool_call", e.ToolCall)
-}
-
-func (tr ToolResult) MarshalZerologObject(ev *zerolog.Event) {
-	ev.Str("id", tr.ID).Str("result", tr.Result)
-	if tr.Name != "" {
-		ev.Str("name", tr.Name)
-	}
-}
-
-func (e EventToolResult) MarshalZerologObject(ev *zerolog.Event) {
-	e.EventImpl.MarshalZerologObject(ev)
-	ev.Object("tool_result", e.ToolResult)
-}
-
-func (e EventPartialCompletion) MarshalZerologObject(ev *zerolog.Event) {
-	e.EventImpl.MarshalZerologObject(ev)
-	ev.Str("delta", e.Delta).Str("completion", e.Completion)
-}
-
-// EventLog represents a generic log record emitted during inference (by engine, middleware or tools)
-type EventLog struct {
-	EventImpl
-	Level   string                 `json:"level"`
-	Message string                 `json:"message"`
-	Fields  map[string]interface{} `json:"fields,omitempty"`
-}
-
-func NewLogEvent(metadata EventMetadata, level string, message string, fields map[string]interface{}) *EventLog {
-	return &EventLog{
-		EventImpl: EventImpl{
-			Type_:     EventTypeLog,
-			Metadata_: metadata,
-			payload:   nil,
-		},
-		Level:   level,
-		Message: message,
-		Fields:  fields,
-	}
-}
-
-var _ Event = &EventLog{}
-
-func (e EventLog) MarshalZerologObject(ev *zerolog.Event) {
-	e.EventImpl.MarshalZerologObject(ev)
-	ev.Str("level", e.Level).Str("message", e.Message)
-	if len(e.Fields) > 0 {
-		ev.Dict("fields", zerolog.Dict().Fields(e.Fields))
-	}
-}
-
-// EventInfo is a lightweight informational message for user-facing notifications
-type EventInfo struct {
-	EventImpl
-	Message string                 `json:"message"`
-	Data    map[string]interface{} `json:"data,omitempty"`
-}
-
-func NewInfoEvent(metadata EventMetadata, message string, data map[string]interface{}) *EventInfo {
-	return &EventInfo{
-		EventImpl: EventImpl{
-			Type_:     EventTypeInfo,
-			Metadata_: metadata,
-			payload:   nil,
-		},
-		Message: message,
-		Data:    data,
-	}
-}
-
-var _ Event = &EventInfo{}
-
-func (e EventInfo) MarshalZerologObject(ev *zerolog.Event) {
-	e.EventImpl.MarshalZerologObject(ev)
-	ev.Str("message", e.Message)
-	if len(e.Data) > 0 {
-		ev.Dict("data", zerolog.Dict().Fields(e.Data))
-	}
-}
-
-// EventAgentModeSwitch: exported custom event with analysis and new mode
-// Message carries a short title; Data should include "from", "to", and optionally "analysis"
-type EventAgentModeSwitch struct {
-	EventImpl
-	Message string                 `json:"message"`
-	Data    map[string]interface{} `json:"data,omitempty"`
-}
-
-func NewAgentModeSwitchEvent(metadata EventMetadata, from string, to string, analysis string) *EventAgentModeSwitch {
-	data := map[string]interface{}{"from": from, "to": to}
-	if analysis != "" {
-		data["analysis"] = analysis
-	}
-	return &EventAgentModeSwitch{
-		EventImpl: EventImpl{Type_: EventTypeAgentModeSwitch, Metadata_: metadata},
-		Message:   "agentmode: mode switched",
-		Data:      data,
-	}
-}
-
-var _ Event = &EventAgentModeSwitch{}
-
-// --- Web search custom events ---
-
-type EventWebSearchStarted struct {
-	EventImpl
-	ItemID string `json:"item_id,omitempty"`
-	Query  string `json:"query,omitempty"`
-}
-
-func NewWebSearchStarted(metadata EventMetadata, itemID, query string) *EventWebSearchStarted {
-	return &EventWebSearchStarted{EventImpl: EventImpl{Type_: EventTypeWebSearchStarted, Metadata_: metadata}, ItemID: itemID, Query: query}
-}
-
-type EventWebSearchSearching struct {
-	EventImpl
-	ItemID string `json:"item_id,omitempty"`
-}
-
-func NewWebSearchSearching(metadata EventMetadata, itemID string) *EventWebSearchSearching {
-	return &EventWebSearchSearching{EventImpl: EventImpl{Type_: EventTypeWebSearchSearching, Metadata_: metadata}, ItemID: itemID}
-}
-
-type EventWebSearchOpenPage struct {
-	EventImpl
-	ItemID string `json:"item_id,omitempty"`
-	URL    string `json:"url,omitempty"`
-}
-
-func NewWebSearchOpenPage(metadata EventMetadata, itemID, url string) *EventWebSearchOpenPage {
-	return &EventWebSearchOpenPage{EventImpl: EventImpl{Type_: EventTypeWebSearchOpenPage, Metadata_: metadata}, ItemID: itemID, URL: url}
-}
-
-type EventWebSearchDone struct {
-	EventImpl
-	ItemID string `json:"item_id,omitempty"`
-}
-
-func NewWebSearchDone(metadata EventMetadata, itemID string) *EventWebSearchDone {
-	return &EventWebSearchDone{EventImpl: EventImpl{Type_: EventTypeWebSearchDone, Metadata_: metadata}, ItemID: itemID}
-}
-
-// Citation event attached to streamed output text
-type EventCitation struct {
-	EventImpl
-	Title           string `json:"title,omitempty"`
-	URL             string `json:"url,omitempty"`
-	StartIndex      *int   `json:"start_index,omitempty"`
-	EndIndex        *int   `json:"end_index,omitempty"`
-	OutputIndex     *int   `json:"output_index,omitempty"`
-	ContentIndex    *int   `json:"content_index,omitempty"`
-	AnnotationIndex *int   `json:"annotation_index,omitempty"`
-}
-
-func NewCitation(metadata EventMetadata, title, url string, start, end, outputIdx, contentIdx, annIdx *int) *EventCitation {
-	return &EventCitation{EventImpl: EventImpl{Type_: EventTypeCitation, Metadata_: metadata}, Title: title, URL: url, StartIndex: start, EndIndex: end, OutputIndex: outputIdx, ContentIndex: contentIdx, AnnotationIndex: annIdx}
-}
-
-// File search custom events
-type EventFileSearchStarted struct {
-	EventImpl
-	ItemID string `json:"item_id,omitempty"`
-}
-
-func NewFileSearchStarted(metadata EventMetadata, itemID string) *EventFileSearchStarted {
-	return &EventFileSearchStarted{EventImpl: EventImpl{Type_: EventTypeFileSearchStarted, Metadata_: metadata}, ItemID: itemID}
-}
-
-type EventFileSearchSearching struct {
-	EventImpl
-	ItemID string `json:"item_id,omitempty"`
-}
-
-func NewFileSearchSearching(metadata EventMetadata, itemID string) *EventFileSearchSearching {
-	return &EventFileSearchSearching{EventImpl: EventImpl{Type_: EventTypeFileSearchSearching, Metadata_: metadata}, ItemID: itemID}
-}
-
-type EventFileSearchDone struct {
-	EventImpl
-	ItemID string `json:"item_id,omitempty"`
-}
-
-func NewFileSearchDone(metadata EventMetadata, itemID string) *EventFileSearchDone {
-	return &EventFileSearchDone{EventImpl: EventImpl{Type_: EventTypeFileSearchDone, Metadata_: metadata}, ItemID: itemID}
-}
-
-// Code interpreter custom events
-type EventCodeInterpreterStarted struct {
-	EventImpl
-	ItemID string `json:"item_id,omitempty"`
-}
-
-func NewCodeInterpreterStarted(metadata EventMetadata, itemID string) *EventCodeInterpreterStarted {
-	return &EventCodeInterpreterStarted{EventImpl: EventImpl{Type_: EventTypeCodeInterpreterStarted, Metadata_: metadata}, ItemID: itemID}
-}
-
-type EventCodeInterpreterInterpreting struct {
-	EventImpl
-	ItemID string `json:"item_id,omitempty"`
-}
-
-func NewCodeInterpreterInterpreting(metadata EventMetadata, itemID string) *EventCodeInterpreterInterpreting {
-	return &EventCodeInterpreterInterpreting{EventImpl: EventImpl{Type_: EventTypeCodeInterpreterInterpreting, Metadata_: metadata}, ItemID: itemID}
-}
-
-type EventCodeInterpreterDone struct {
-	EventImpl
-	ItemID string `json:"item_id,omitempty"`
-}
-
-func NewCodeInterpreterDone(metadata EventMetadata, itemID string) *EventCodeInterpreterDone {
-	return &EventCodeInterpreterDone{EventImpl: EventImpl{Type_: EventTypeCodeInterpreterDone, Metadata_: metadata}, ItemID: itemID}
-}
-
-type EventCodeInterpreterCodeDelta struct {
-	EventImpl
-	ItemID string `json:"item_id,omitempty"`
-	Delta  string `json:"delta"`
-}
-
-func NewCodeInterpreterCodeDelta(metadata EventMetadata, itemID, delta string) *EventCodeInterpreterCodeDelta {
-	return &EventCodeInterpreterCodeDelta{EventImpl: EventImpl{Type_: EventTypeCodeInterpreterCodeDelta, Metadata_: metadata}, ItemID: itemID, Delta: delta}
-}
-
-type EventCodeInterpreterCodeDone struct {
-	EventImpl
-	ItemID string `json:"item_id,omitempty"`
-	Code   string `json:"code"`
-}
-
-func NewCodeInterpreterCodeDone(metadata EventMetadata, itemID, code string) *EventCodeInterpreterCodeDone {
-	return &EventCodeInterpreterCodeDone{EventImpl: EventImpl{Type_: EventTypeCodeInterpreterCodeDone, Metadata_: metadata}, ItemID: itemID, Code: code}
-}
-
-// MCP
-type EventMCPArgsDelta struct {
-	EventImpl
-	ItemID string `json:"item_id,omitempty"`
-	Delta  string `json:"delta"`
-}
-
-func NewMCPArgsDelta(metadata EventMetadata, itemID, delta string) *EventMCPArgsDelta {
-	return &EventMCPArgsDelta{EventImpl: EventImpl{Type_: EventTypeMCPArgsDelta, Metadata_: metadata}, ItemID: itemID, Delta: delta}
-}
-
-type EventMCPArgsDone struct {
-	EventImpl
-	ItemID    string `json:"item_id,omitempty"`
-	Arguments string `json:"arguments"`
-}
-
-func NewMCPArgsDone(metadata EventMetadata, itemID, args string) *EventMCPArgsDone {
-	return &EventMCPArgsDone{EventImpl: EventImpl{Type_: EventTypeMCPArgsDone, Metadata_: metadata}, ItemID: itemID, Arguments: args}
-}
-
-type EventMCPInProgress struct {
-	EventImpl
-	ItemID string `json:"item_id,omitempty"`
-}
-
-func NewMCPInProgress(metadata EventMetadata, itemID string) *EventMCPInProgress {
-	return &EventMCPInProgress{EventImpl: EventImpl{Type_: EventTypeMCPInProgress, Metadata_: metadata}, ItemID: itemID}
-}
-
-type EventMCPCompleted struct {
-	EventImpl
-	ItemID string `json:"item_id,omitempty"`
-}
-
-func NewMCPCompleted(metadata EventMetadata, itemID string) *EventMCPCompleted {
-	return &EventMCPCompleted{EventImpl: EventImpl{Type_: EventTypeMCPCompleted, Metadata_: metadata}, ItemID: itemID}
-}
-
-type EventMCPFailed struct {
-	EventImpl
-	ItemID string `json:"item_id,omitempty"`
-}
-
-func NewMCPFailed(metadata EventMetadata, itemID string) *EventMCPFailed {
-	return &EventMCPFailed{EventImpl: EventImpl{Type_: EventTypeMCPFailed, Metadata_: metadata}, ItemID: itemID}
-}
-
-type EventMCPListInProgress struct {
-	EventImpl
-	ItemID string `json:"item_id,omitempty"`
-}
-
-func NewMCPListInProgress(metadata EventMetadata, itemID string) *EventMCPListInProgress {
-	return &EventMCPListInProgress{EventImpl: EventImpl{Type_: EventTypeMCPListInProgress, Metadata_: metadata}, ItemID: itemID}
-}
-
-type EventMCPListCompleted struct {
-	EventImpl
-	ItemID string `json:"item_id,omitempty"`
-}
-
-func NewMCPListCompleted(metadata EventMetadata, itemID string) *EventMCPListCompleted {
-	return &EventMCPListCompleted{EventImpl: EventImpl{Type_: EventTypeMCPListCompleted, Metadata_: metadata}, ItemID: itemID}
-}
-
-type EventMCPListFailed struct {
-	EventImpl
-	ItemID string `json:"item_id,omitempty"`
-}
-
-func NewMCPListFailed(metadata EventMetadata, itemID string) *EventMCPListFailed {
-	return &EventMCPListFailed{EventImpl: EventImpl{Type_: EventTypeMCPListFailed, Metadata_: metadata}, ItemID: itemID}
-}
-
-// Image generation
-type EventImageGenInProgress struct {
-	EventImpl
-	ItemID string `json:"item_id,omitempty"`
-}
-
-func NewImageGenInProgress(metadata EventMetadata, itemID string) *EventImageGenInProgress {
-	return &EventImageGenInProgress{EventImpl: EventImpl{Type_: EventTypeImageGenInProgress, Metadata_: metadata}, ItemID: itemID}
-}
-
-type EventImageGenGenerating struct {
-	EventImpl
-	ItemID string `json:"item_id,omitempty"`
-}
-
-func NewImageGenGenerating(metadata EventMetadata, itemID string) *EventImageGenGenerating {
-	return &EventImageGenGenerating{EventImpl: EventImpl{Type_: EventTypeImageGenGenerating, Metadata_: metadata}, ItemID: itemID}
-}
-
-type EventImageGenPartialImage struct {
-	EventImpl
-	ItemID             string `json:"item_id,omitempty"`
-	PartialImageBase64 string `json:"partial_image_base64,omitempty"`
-}
-
-func NewImageGenPartialImage(metadata EventMetadata, itemID, b64 string) *EventImageGenPartialImage {
-	return &EventImageGenPartialImage{EventImpl: EventImpl{Type_: EventTypeImageGenPartialImage, Metadata_: metadata}, ItemID: itemID, PartialImageBase64: b64}
-}
-
-type EventImageGenCompleted struct {
-	EventImpl
-	ItemID string `json:"item_id,omitempty"`
-}
-
-func NewImageGenCompleted(metadata EventMetadata, itemID string) *EventImageGenCompleted {
-	return &EventImageGenCompleted{EventImpl: EventImpl{Type_: EventTypeImageGenCompleted, Metadata_: metadata}, ItemID: itemID}
-}
-
-// Normalized results
-type SearchResult struct {
-	URL        string         `json:"url,omitempty"`
-	Title      string         `json:"title,omitempty"`
-	Snippet    string         `json:"snippet,omitempty"`
-	Extensions map[string]any `json:"ext,omitempty"`
-}
-type EventToolSearchResults struct {
-	EventImpl
-	Tool    string         `json:"tool"`
-	ItemID  string         `json:"item_id,omitempty"`
-	Results []SearchResult `json:"results"`
-}
-
-func NewToolSearchResults(metadata EventMetadata, tool, itemID string, res []SearchResult) *EventToolSearchResults {
-	return &EventToolSearchResults{EventImpl: EventImpl{Type_: EventTypeToolSearchResults, Metadata_: metadata}, Tool: tool, ItemID: itemID, Results: res}
 }
